crm/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.9.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
Django==4.2.3
django-restframework==0.0.1
djangorestframework==3.14.0
idna==3.4
joblib==1.3.1
nltk==3.8.1
psycopg2==2.9.6
pytz==2023.3
regex==2023.6.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
textblob==0.15.3
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.3
